+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- DF-1144 Made DELETE behavior consistent across local and all remote file services
 
 ## [0.3.0] - 2017-06-05
 ### Changed